+++ libs/safe/src/lib/components/grid-layout/edit-layout-modal/public-api.ts
@@ -1,2 +1,1 @@
 export * from './edit-layout-modal.component';
-export * from './edit-layout-modal.module';