+++ libs/ui/src/index.ts
@@ -10,6 +10,9 @@ export * from './lib/select/select.module';
 export * from './lib/spinner/spinner.module';
 export * from './lib/toggle/toggle.module';
 export * from './lib/breadcrumbs/breadcrumbs.module';
+export * from './lib/avatar-group/avatar-group.module';
+export * from './lib/breadcrumbs/breadcrumbs.module';
+export * from './lib/divider/divider.module';
 
 // === ENUMS === //
 export * from './lib/shared/size.enum';
@@ -25,5 +28,4 @@ export * from './lib/breadcrumbs/enums/breadcrumb-separator.enum';
 // === INTERFACES === //
 export * from './lib/toggle/interfaces/toggle-label.interface';
 export * from './lib/toggle/interfaces/toggle-icon.interface';
-export * from './lib/divider/divider.module';
 export * from './lib/breadcrumbs/interfaces/breadcrumb.interface';