+++ libs/ui/src/lib/ui.module.ts
@@ -7,6 +7,7 @@ import { FlyoutMenuModule } from './flyout-menu/flyout-menu.module';
 import { IconModule } from './icon/icon.module';
 import { SelectModule } from './select/select.module';
 import { SpinnerModule } from './spinner/spinner.module';
+import { AvatarGroupModule } from './avatar-group/avatar-group.module';
 import { ToggleModule } from './toggle/toggle.module';
 import { DividerModule } from './divider/divider.module';
 
@@ -18,6 +19,7 @@ import { DividerModule } from './divider/divider.module';
   exports: [
     AvatarModule,
     ButtonModule,
+    AvatarGroupModule,
     CheckboxModule,
     FlyoutMenuModule,
     IconModule,